Renamed the nl.soccar.util package to nl.soccar.ui.util.
